rename secrets manager utils
