+++ packages/aws-client/SecretsManager.js
